api/requirements.txt
# ==================================================

# Backend – Python Requirements (requirements.txt)

# ==================================================

```txt
# ==================================================
# Backend Framework
# ==================================================
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
python-dotenv==1.0.1

# ==================================================
# Database
# ==================================================
pymongo==4.6.1

# ==================================================
# Machine Learning – Severity Prediction
# ==================================================
scikit-learn==1.4.0
xgboost==2.0.3

# ==================================================
# Hotspot Detection – KDE & Clustering
# ==================================================
hdbscan==0.8.33
scipy==1.12.0

# ==================================================
# Data Processing
# ==================================================
numpy==1.26.4
pandas==2.2.0

# ==================================================
# Model Persistence
# ==================================================
joblib==1.3.2
```

# ==================================================

# Frontend – npm Dependencies

# ==================================================

```txt
react
react-dom
vite
axios
leaflet
react-router-dom
```
